+++ src/dashboardWebView/DashboardCommand.ts
@@ -1,9 +1,11 @@
 export enum DashboardCommand {
   loading = "loading",
   pages = "pages",
+  searchPages = "searchPages",
   settings = "settings",
   media = "media",
   viewData = "viewData",
   mediaUpdate = "mediaUpdate",
-  dataFileEntries = "dataFileEntries"
+  dataFileEntries = "dataFileEntries",
+  searchReady = "searchReady",
 }