+++ src/dashboardWebView/components/Header/Searchbox.tsx
@@ -1,8 +1,8 @@
-import {SearchIcon} from '@heroicons/react/solid';
+import {SearchIcon, XCircleIcon} from '@heroicons/react/solid';
 import * as React from 'react';
-import { useRecoilState } from 'recoil';
+import { useRecoilState, useRecoilValue } from 'recoil';
 import { useDebounce } from '../../../hooks/useDebounce';
-import { SearchAtom } from '../../state';
+import { SearchAtom, SearchReadyAtom } from '../../state';
 import { RefreshPages } from './RefreshPages';
 
 export interface ISearchboxProps {
@@ -12,11 +12,17 @@ export interface ISearchboxProps {
 export const Searchbox: React.FunctionComponent<ISearchboxProps> = ({placeholder}: React.PropsWithChildren<ISearchboxProps>) => {
   const [ value, setValue ] = React.useState('');
   const [ debounceSearchValue, setDebounceValue ] = useRecoilState(SearchAtom);
+  const searchReady = useRecoilValue(SearchReadyAtom);
   const debounceSearch = useDebounce<string>(value, 500);
 
   const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
     setValue(event.target.value);
   };
+
+  const reset = React.useCallback(() => {
+    setValue('');
+    setDebounceValue('');
+  }, [setValue, setDebounceValue]);
 
   React.useEffect(() => {    
     if (!debounceSearchValue && value) {
@@ -38,13 +44,22 @@ export const Searchbox: React.FunctionComponent<ISearchboxProps> = ({placeholder
           </div>
 
           <input
-            type="search"
+            type="text"
             name="search"
-            className={`block w-full py-2 pl-10 pr-3 sm:text-sm bg-white dark:bg-vulcan-300 border border-gray-300 dark:border-vulcan-100 text-vulcan-500 dark:text-whisper-500 placeholder-gray-400 dark:placeholder-whisper-800 focus:outline-none`}
+            className={`block w-full py-2 pl-10 pr-3 sm:text-sm bg-white dark:bg-vulcan-300 border border-gray-300 dark:border-vulcan-100 text-vulcan-500 dark:text-whisper-500 placeholder-gray-400 dark:placeholder-whisper-800 focus:outline-none appearance-none disabled:opacity-50`}
             placeholder={placeholder || "Search"}
             value={value}
             onChange={handleChange}
+            disabled={!searchReady}
           />
+
+          {
+            value && (
+              <button onClick={reset} className="absolute inset-y-0 right-0 pr-3 flex items-center">
+                <XCircleIcon className="h-5 w-5 text-gray-400" aria-hidden="true" />
+              </button>
+            )
+          }
         </div>
       </div>