+++ src/dashboardWebView/hooks/useMessages.tsx
@@ -3,7 +3,7 @@ import { useRecoilState } from 'recoil';
 import { DashboardCommand } from '../DashboardCommand';
 import { DashboardMessage } from '../DashboardMessage';
 import { Page } from '../models/Page';
-import { DashboardViewAtom, LoadingAtom, SettingsAtom, ViewDataAtom } from '../state';
+import { DashboardViewAtom, LoadingAtom, SettingsAtom, ViewDataAtom, SearchReadyAtom } from '../state';
 import { Messenger } from '@estruyf/vscode/dist/client';
 import { EventData } from '@estruyf/vscode/dist/models';
 import { NavigationType } from '../models';
@@ -14,6 +14,7 @@ export default function useMessages() {
   const [settings, setSettings] = useRecoilState(SettingsAtom);
   const [viewData, setViewData] = useRecoilState(ViewDataAtom);
   const [, setView] = useRecoilState(DashboardViewAtom);
+  const [, setSearchReady] = useRecoilState(SearchReadyAtom);
 
   Messenger.listen((message: MessageEvent<EventData<any>>) => {
     switch (message.data.command) {
@@ -38,6 +39,9 @@ export default function useMessages() {
       case DashboardCommand.pages:
         setPages(message.data.data);
         setLoading(false);
+        break;
+      case DashboardCommand.searchReady:
+        setSearchReady(true);
         break;
     }
   });