+++ src/dashboardWebView/state/atom/SearchReadyAtom.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const SearchReadyAtom = atom<boolean>({
+  key: 'SearchReadyAtom',
+  default: false
+});