+++ src/dashboardWebView/state/atom/index.ts
@@ -8,6 +8,7 @@ export * from './MediaFoldersAtom';
 export * from './MediaTotalAtom';
 export * from './PageAtom';
 export * from './SearchAtom';
+export * from './SearchReadyAtom';
 export * from './SelectedMediaFolderAtom';
 export * from './SettingsAtom';
 export * from './SortingAtom';